Clean build warnings for c++ (Raspbian 6.3.0-18+rpi1+deb9u1) 6.3.0 20170516

+++ ili9341.cpp
@@ -13,7 +13,7 @@ static void ILI9341ClearScreen()
   for(int y = 0; y < DISPLAY_HEIGHT; ++y)
   {
     SPI_TRANSFER(DISPLAY_SET_CURSOR_X, 0, 0, DISPLAY_WIDTH >> 8, DISPLAY_WIDTH & 0xFF);
-    SPI_TRANSFER(DISPLAY_SET_CURSOR_Y, y >> 8, y & 0xFF, DISPLAY_HEIGHT >> 8, DISPLAY_HEIGHT & 0xFF);
+    SPI_TRANSFER(DISPLAY_SET_CURSOR_Y, (uint8_t)(y >> 8), y & 0xFF, DISPLAY_HEIGHT >> 8, DISPLAY_HEIGHT & 0xFF);
     SPITask *clearLine = AllocTask(DISPLAY_WIDTH*2);
     clearLine->cmd = DISPLAY_WRITE_PIXELS;
     memset(clearLine->data, 0, clearLine->size);
